% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keydur.R
\name{keydur}
\alias{keydur}
\title{keyrate duration function}
\usage{
keydur(keyrates, ttm, coupon, yield, freq, targetdur = NULL)
}
\arguments{
\item{keyrates}{list of keyrates}

\item{ttm}{time until maturity in years}

\item{coupon}{in percent}

\item{yield}{in percent}

\item{freq}{frequency of coupon payments}

\item{targetdur}{optional argument what the sum of keyrate durations has to be}
}
\value{
tibble of key rate durations
}
\description{
Calculates keyrate duration of a bond
}
\examples{
\dontrun{
keydur(c(2,5,10),6.2,1,2,1)
}
}
\author{
wkapga
}
